==> eda/1.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../data/realestate_data_sample.csv')
df

selected_cols = ['price', 'details.dien-tich-su-dung', 'details.phong-ngu', 'details.nha-tam', 'district']

df = df[selected_cols]

def extract_used_area(area):
    try:
        area = area.replace("m2", "")
        area = area.replace(",", ".")

        area = area.strip()
        try:
            return float(area)
        except:
            return np.nan
    except:
        return np.nan
df['used_area'] = df['details.dien-tich-su-dung'].apply(extract_used_area)

df = df.dropna()

district = list(df['district'].unique())
district_dict = dict(zip(district, range(len(district))))
district_dict

df['district'] = df['district'].map(district_dict)
df['district'].value_counts()

df = df[(df['district'] != 23) & (df['district'] != 20)]

df = df.reset_index(drop = True)

df = df.rename(columns = {
    'details.phong-ngu': 'num_of_bedroom',
    'details.nha-tam': 'num_of_bathroom',
})
df = df[['num_of_bedroom', 'num_of_bathroom', 'used_area', 'district', 'price']]

df.to_parquet('../data/processed_realestate_data_sample.parquet', index = False)

df

df['target'] = df['price'] * 1000 / df['used_area']

feat_set1 = ['num_of_bedroom', 'district']
feat_set2 = ['num_of_bathroom', 'district']
feat_set3 = ['num_of_bathroom', 'district', 'num_of_bedroom']
feat_set4 = ['num_of_bathroom', 'district', 'num_of_bedroom', 'used_area']


data_df1 = df[feat_set1]
data_df2 = df[feat_set2]
data_df3 = df[feat_set3]
data_df4 = df[feat_set4]
target_df = pd.DataFrame(data=df.target, columns=["target"])

timestamps = pd.date_range(
    end=pd.Timestamp.now(),
    periods=len(df),
    freq='D').to_frame(name="event_timestamp", index=False)

data_df1 = pd.concat(objs=[data_df1, timestamps], axis=1)
data_df2 = pd.concat(objs=[data_df2, timestamps], axis=1)
data_df3 = pd.concat(objs=[data_df3, timestamps], axis=1)
data_df4 = pd.concat(objs=[data_df4, timestamps], axis=1)
target_df = pd.concat(objs=[target_df, timestamps], axis=1)

realestate_ids = pd.DataFrame(data=list(range(len(df))), columns=["realestate_id"])

data_df1 = pd.concat(objs=[data_df1, realestate_ids], axis=1)
data_df2 = pd.concat(objs=[data_df2, realestate_ids], axis=1)
data_df3 = pd.concat(objs=[data_df3, realestate_ids], axis=1)
data_df4 = pd.concat(objs=[data_df4, realestate_ids], axis=1)
target_df = pd.concat(objs=[target_df, realestate_ids], axis=1)

data_df1.to_parquet(path='../data/data_df1.parquet')
data_df2.to_parquet(path='../data/data_df2.parquet')
data_df3.to_parquet(path='../data/data_df3.parquet')
data_df4.to_parquet(path='../data/data_df4.parquet')
target_df.to_parquet(path='../data/target_df.parquet')

data_df4['event_timestamp']
