==> eda/3.ipynb <==
import json

data = json.load(open('../data/process_v1/realestate-dataset.json', 'r'))
data[0]


data[0]['mediaInfo']['certificateOfLandUseRight']['certificateStatus']

def get_compact_info(item):
    numberOfFloors = item['houseInfo']['value']['numberOfFloors']
    numberOfBathRooms = item['houseInfo']['value']['numberOfBathRooms']
    numberOfKitchens = item['houseInfo']['value']['numberOfKitchens']
    numberOfLivingRooms = item['houseInfo']['value']['numberOfLivingRooms']
    numberOfGarages = item['houseInfo']['value']['numberOfGarages']
    ward = item['propertyBasicInfo']['address']['value']['ward']
    street = item['propertyBasicInfo']['address']['value']['street']
    district = item['propertyBasicInfo']['address']['value']['district']
    city = item['propertyBasicInfo']['address']['value']['city']
    lat = item['propertyBasicInfo']['geolocation']['value']['latitude']['value']
    lon = item['propertyBasicInfo']['geolocation']['value']['longitude']['value']
    typeOfRealEstate = item['propertyBasicInfo']['typeOfRealEstate']['value']
    frontWidth = item['propertyBasicInfo']['frontWidth']['value']
    endWidth = item['propertyBasicInfo']['endWidth']['value']
    facade = item['propertyBasicInfo']['facade']['value']
    houseDirection = item['propertyBasicInfo']['houseDirection']['value']
    landSize = item['propertyBasicInfo']['landSize']['value']
    price = item['propertyBasicInfo']['price']['value']
    unitPrice = item['propertyBasicInfo']['unitPrice']['value']
    distanceToNearestRoad = item['propertyBasicInfo']['distanceToNearestRoad']['value']
    frontRoadWidth = item['propertyBasicInfo']['frontRoadWidth']['value']
    accessibility = item['propertyBasicInfo']['accessibility']['value']
    landType = item['propertyBasicInfo']['landType']['value']
    description = item['propertyBasicInfo']['description']['value']
    time = item['crawlInfo']['time']

    return {
        "numberOfFloors": numberOfFloors,
        "ward": ward,
        "street": street,
        "district": district,
        "city": city,
        "lat": lat,
        "lon": lon,
        "typeOfRealEstate": typeOfRealEstate,
        "frontWidth": frontWidth,
        "endWidth": endWidth,
        "facade": facade,
        "houseDirection": houseDirection,
        "landSize": landSize,
        "price": price,
        "unitPrice": unitPrice,
        "distanceToNearestRoad": distanceToNearestRoad,
        "frontRoadWidth": frontRoadWidth,
        "accessibility": accessibility,
        "landType": landType,
        "description": description,
        'time': time
    }

get_compact_info(data[9000])
