==> eda/5.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../data/process_v1/process_data_2.csv')

df.head(2)

df['endWidth'] = df['endWidth'].replace(0, np.nan)
df['frontRoadWidth'] = df['frontRoadWidth'].replace(0, np.nan)
df['numberOfFloors'] = df['numberOfFloors'].replace(0, np.nan)

df['numberOfBathRooms'] = df['numberOfBathRooms'].replace(0, np.nan)
df['numberOfBedRooms'] = df['numberOfBedRooms'].replace(0, np.nan)
df['numberOfKitchens'] = df['numberOfKitchens'].replace(0, np.nan)
df['numberOfGarages'] = df['numberOfGarages'].replace(0, np.nan)
df['certificateOfLandUseRight'] == (df['certificateOfLandUseRight'] == 'yes').astype(np.int32)



df['numberOfFloors'].value_counts()

df = df[df['numberOfFloors'] <= 100]
df

del df['frontRoadWidth']

del df['distanceToNearestRoad']

for c in df.columns.tolist():
    print(c, df[c].isnull().sum() / df.shape[0])

df = df.reset_index(drop = True)
display(df)
df.to_csv('../data/process_v1/process_data_3.csv', index = False)
