==> eda/6.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../data/process_v1/process_data_3.csv')

df

df = df[df['street'].isnull() == False]

def concat(a, b):
    return f'{a} - {b}'
df['ward'] = df.apply(lambda x: concat(x['district'], x['ward']), axis = 1)
df['street'] = df.apply(lambda x: concat(x['ward'], x['street']), axis = 1)

item = list(df[df['city'] == 'hồ chí minh'].ward.unique())
import json

dictionary = [i for i in item if i != '']

with open("./test.json", "w", encoding='utf8') as outfile:
    json.dump(dictionary, outfile, ensure_ascii=False)

df = df.reset_index(drop = True)
display(df)
df.to_csv('../data/process_v1/process_data_4.csv', index = False)
