==> eda/7.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../data/process_v1/process_data_4.csv')

del df['unitPrice']

del df['landType']

df.head(3)

df['w'] = df['frontWidth']
df['h'] = df['landSize'] / df['w']

df = df.replace(np.inf, np.nan)

df.columns.tolist()

location_df = df[['lat', 'lon']].drop_duplicates()

facility_500_df = pd.read_csv('../feature_repo/src/feature_engineering/facility/data/table/hcm_hn_distance_500_facility_count.csv')
facility_1000_df = pd.read_csv('../feature_repo/src/feature_engineering/facility/data/table/hcm_hn_distance_1000_facility_count.csv')
facility_2000_df = pd.read_csv('../feature_repo/src/feature_engineering/facility/data/table/hcm_hn_distance_2000_facility_count.csv')

def rename_col(dataframe, distance):
    cols = [c for c in dataframe.columns if c  not in ['lat', 'lon']]
    rename_cols = [f'num_of_{col}_in_{distance}m_radius' for col in cols]

    rename_col_dict = dict(zip(cols, rename_cols))
    return dataframe.rename(columns = rename_col_dict)

facility_500_df = rename_col(facility_500_df, 500)
facility_1000_df = rename_col(facility_1000_df, 1000)
facility_2000_df = rename_col(facility_2000_df, 2000)


for facility_df in [facility_500_df, facility_1000_df, facility_2000_df]:
    df = df.merge(facility_df, how='left', on = ['lat', 'lon'])

df

num_cols = []
cat_cols = []
for c in df.columns.tolist():
    if df[c].dtype != 'object':
        num_cols.append(c)
    else:
        cat_cols.append(c)


cat_cols

num_cols

df[
'typeOfRealEstate'
].value_counts()

certificateOfLandUseRight_list = list(df['certificateOfLandUseRight'].unique())
certificateOfLandUseRight_dict = dict(zip(certificateOfLandUseRight_list, range(len(certificateOfLandUseRight_list))))
df['certificateOfLandUseRight'] = df['certificateOfLandUseRight'].map(certificateOfLandUseRight_dict)

ward_list = list(df['ward'].unique())
ward_dict = dict(zip(ward_list, range(len(ward_list))))
df['ward'] = df['ward'].map(ward_dict)
street_list = list(df['street'].unique())
street_dict = dict(zip(street_list, range(len(street_list))))
df['street'] = df['street'].map(street_dict)
district_list = list(df['district'].unique())
district_dict = dict(zip(district_list, range(len(district_list))))
df['district'] = df['district'].map(district_dict)
city_list = list(df['city'].unique())
city_dict = dict(zip(city_list, range(len(city_list))))
df['city'] = df['city'].map(city_dict)
typeOfRealEstate_list = list(df['typeOfRealEstate'].unique())
typeOfRealEstate_dict = dict(zip(typeOfRealEstate_list, range(len(typeOfRealEstate_list))))
df['typeOfRealEstate'] = df['typeOfRealEstate'].map(typeOfRealEstate_dict)
facade_list = list(df['facade'].unique())
facade_dict = dict(zip(facade_list, range(len(facade_list))))
df['facade'] = df['facade'].map(facade_dict)
houseDirection_list = list(df['houseDirection'].unique())
houseDirection_dict = dict(zip(houseDirection_list, range(len(houseDirection_list))))
df['houseDirection'] = df['houseDirection'].map(houseDirection_dict)
accessibility_list = list(df['accessibility'].unique())
accessibility_dict = dict(zip(accessibility_list, range(len(accessibility_list))))
df['accessibility'] = df['accessibility'].map(accessibility_dict)

ward_dict

# {'thủ đức - tam phú': 0,
#  'quận 7 - tân phong': 1,
#  'nam từ liêm - mễ trì': 2,
#  'hai bà trưng - minh khai': 3,
#  'suburb_north - bắc sơn': 4,
#  'quận 7 - tân quy': 5,
#  'cầu giấy - mai dịch': 6,
#  'long biên - bồ đề': 7,
#  'đống đa - thịnh quang': 8,
#  'hoài đức - an thượng': 9,
#  'gia lâm - trâu quỳ': 10,
#  'tân bình - phường 12': 11,
#  'hai bà trưng - thanh nhàn': 12,
#  'quận 1 - tân định': 13,
#  'nam từ liêm - trung văn': 14,
#  'hà đông - văn quán': 15,
#  'hoài đức - vân canh': 16,
#  'thủ đức - hiệp bình phước': 17,
#  'long biên - gia thụy': 18,
#  'long biên - đức giang': 19,
#  'quận 12 - tân thới nhất': 20,
#  'phú nhuận - phường 1': 21,
#  'quận 9 - long thạnh mỹ': 22,
#  'hai bà trưng - thanh lương': 23,
#  'quận 11 - phường 5': 24,

city_dict

df['accessibility'].value_counts()

accessibility_dict

import json

dictionary = accessibility_dict

with open("./test.json", "w", encoding='utf8') as outfile:
    json.dump(dictionary, outfile, ensure_ascii=False)

df1 = pd.read_csv('../feature_repo/src/feature_engineering/population/data/table/hcm_population.csv')
df2 = pd.read_csv('../feature_repo/src/feature_engineering/population/data/table/hn_population.csv')

population_df = pd.concat([
    df1, df2
])

def convert_district_to_num(district):
    if district in ['mê linh',
        'ba vì',
        'phúc thọ',
        'thạch thất',
        'mỹ đức',
        'sơn tây',
        'quốc oai',
        'quốc oai']:
        district = 'suburb_west'
    elif district in ['thanh oai',
'ứng hòa',
'phú xuyên',
'thường tín',
'chương mỹ',]:
        district = 'suburb_south'
    elif district in ['sóc sơn', 'đan phượng']:
        district = 'suburb_north'
    candidate_str = [district, f"quận {district}", f"huyện {district}", f"thị xã {district}"]

    for candidate in candidate_str:
        try:
            return district_dict[candidate]
        except:pass
    print(district)
population_df['district'] = population_df['name'].apply(convert_district_to_num)
population_df

del population_df['average_price']

population_df = population_df.dropna(subset = ['district'])

population_df = population_df.sort_values(by = 'district')
population_df

population_df[['administrative_genre', 'district']]

ad_df = population_df[['administrative_genre', 'district']].drop_duplicates(subset = ['district'], keep = 'first')
ad_df

population_df = population_df.groupby('district')[['acreage', 'population', 'density', 'lat', 'lon', 'num_of_ward']].mean().reset_index()

population_df = ad_df.merge(population_df, how='left', on = ['district'])
population_df

ad_list = list(population_df['administrative_genre'].unique())
ad_dict = dict(zip(ad_list, range(len(ad_list))))
ad_dict

population_df['administrative_genre'] = population_df['administrative_genre'].map(ad_dict)
population_df = population_df.rename(columns = {'lat': 'district_lat', 'lon': 'district_lon'})

population_df

population_df.to_csv('./process_population.csv', index = False)

df = df.merge(population_df, how='left', on = 'district')
df

del df['numberOfKitchens']

del df['numberOfGarages']

city_dict

best_df1 = pd.read_csv('../feature_repo/src/feature_engineering/facility/data/table/hcm_famous_place.csv')
best_df1['city'] = 'hồ chí minh'

best_df2 = pd.read_csv('../feature_repo/src/feature_engineering/facility/data/table/hn_famous_place.csv')
best_df2['city'] = 'hà nội'

best_df = pd.concat([best_df1, best_df2])
best_df

del best_df['address']
best_df['city'] = best_df['city'].map(city_dict)

# df = df.iloc[:2]

import math
from tqdm import tqdm
def distance_func(lat1: float, lon1: float, lat2: float, lon2: float):

    try:
        R = 6371
        dLat = (lat2-lat1) * math.pi / 180
        dLon = (lon2-lon1) * math.pi / 180
        lat1 = lat1 * math.pi / 180
        lat2 = lat2 * math.pi / 180
        a = math.sin(dLat/2) * math.sin(dLat/2) + math.sin(dLon/2) * \
            math.sin(dLon/2) * math.cos(lat1) * math.cos(lat2)
        c = 2 * math.atan2(math.sqrt(a), math.sqrt(1-a))
        d = R * c
        return d*1000
    except:
        print(f"Distance_Func error with: {(lat1, lon1, lat2, lon2)}")

def distance(lat1, lon1, lat2, lon2, city1, city2):
    if city1 != city2:
        return -1
    return distance_func(lat1, lon1, lat2, lon2)


best_arr = best_df.values
for i in tqdm(range(len(best_arr))):
    lat, lon = best_arr[i][0], best_arr[i][1]
    name = best_arr[i][2]
    city = 'hn' if best_arr[i][3] == 0 else 'hcm'

    df[f'distance_{city}_{name}'] = df.apply(lambda x: distance(x['lat'], x['lon'], lat, lon, x['city'], best_arr[i][3]), axis = 1)


df['distance_district'] = df.apply(lambda x: distance_func(x['lat'], x['lon'], x['district_lat'], x['district_lon']), axis = 1)

distance_cols = [c for c in df.columns.tolist() if 'distance' in c]
for c in distance_cols:
    df[c] = np.log(df[c])


distance_cols

df['target'] = df['price'] * 1000 / df['landSize']

df['population'] = np.log(df['population'])

df['density'] = np.log(df['density'])

df['acreage'] = np.log(df['acreage'])

df

df.to_csv('../data/process_v1/process_data_5.csv', index = False)

import numpy as np
import math

math.log(10), np.log(10)
