==> eda/8.ipynb <==
import pandas as pd
import numpy as np
import requests
import re
import nltk
import spacy
from nltk.corpus import stopwords
import string

df = pd.read_csv('../data/process_v1/process_data_4.csv')

df

s1 = u'ÀÁÂÃÈÉÊÌÍÒÓÔÕÙÚÝàáâãèéêìíòóôõùúýĂăĐđĨĩŨũƠơƯưẠạẢảẤấẦầẨẩẪẫẬậẮắẰằẲẳẴẵẶặẸẹẺẻẼẽẾếỀềỂểỄễỆệỈỉỊịỌọỎỏỐốỒồỔổỖỗỘộỚớỜờỞởỠỡỢợỤụỦủỨứỪừỬửỮữỰựỲỳỴỵỶỷỸỹ'
s0 = u'AAAAEEEIIOOOOUUYaaaaeeeiioooouuyAaDdIiUuOoUuAaAaAaAaAaAaAaAaAaAaAaAaEeEeEeEeEeEeEeEeIiIiOoOoOoOoOoOoOoOoOoOoOoOoUuUuUuUuUuUuUuYyYyYyYy'
def remove_accents(input_str):
	s = ''
	for c in input_str:
		if c in s1:
			s += s0[s1.index(c)]
		else:
			s += c
	return s

streets = [remove_accents(item) for item in df['street'].unique()]

address_words = []

for street in streets:
    address_words += street.split(" ")
address_words[0]

eng_1_df = pd.read_csv('../data/process_v1/trans/eng_1_20000.csv')
eng_2_df = pd.read_csv('../data/process_v1/trans/eng_2_20000.csv')
eng_3_df = pd.read_csv('../data/process_v1/trans/eng_3_20000.csv')
eng_4_df = pd.read_csv('../data/process_v1/trans/eng_4_20000.csv')
eng_5_df = pd.read_csv('../data/process_v1/trans/eng_5_20000.csv')
eng_6_df = pd.read_csv('../data/process_v1/trans/eng_6_20000.csv')
eng_7_df = pd.read_csv('../data/process_v1/trans/eng_7_20000.csv')
eng_8_df = pd.read_csv('../data/process_v1/trans/eng_8_30000.csv')
eng_9_df = pd.read_csv('../data/process_v1/trans/eng_9.csv')


eng_df = pd.concat([eng_1_df, eng_2_df, eng_3_df, eng_4_df, eng_5_df,eng_6_df ,eng_7_df, eng_8_df, eng_9_df])
eng_df

eng_df['en_description'] = eng_df['en_description'].fillna("No information")

eng_df["text_lower"] = eng_df["en_description"].str.lower()
eng_df.head()

PUNCT_TO_REMOVE = string.punctuation
def remove_punctuation(text):
    """custom function to remove the punctuation"""
    return text.translate(str.maketrans('', '', PUNCT_TO_REMOVE))

eng_df["text_wo_punct"] = eng_df["text_lower"].apply(lambda text: remove_punctuation(text))
eng_df.head()

STOPWORDS = set(stopwords.words('english'))
def remove_stopwords(text):
    """custom function to remove the stopwords"""
    return " ".join([word for word in str(text).split() if word not in STOPWORDS])

eng_df["text_wo_stop"] = eng_df["text_wo_punct"].apply(lambda text: remove_stopwords(text))
eng_df.head()

from collections import Counter
cnt = Counter()
for text in eng_df["text_wo_stop"].values:
    for word in text.split():
        cnt[word] += 1

FREQWORDS = set([w for (w, wc) in cnt.most_common(10)])
def remove_freqwords(text):
    """custom function to remove the frequent words"""
    return " ".join([word for word in str(text).split() if word not in FREQWORDS])

eng_df["text_wo_stopfreq"] = eng_df["text_wo_stop"].apply(lambda text: remove_freqwords(text))
eng_df.head()

len(cnt.most_common())

n_rare_words = 40000
RAREWORDS = [w for (w, wc) in cnt.most_common()[:-n_rare_words-1:-1]]
NUM_WORDS = [w for (w, wc) in cnt.most_common() if sum([_.isnumeric() for _ in w])]
ADDRESS_WORDS = [w for (w, wc) in cnt.most_common() if w in address_words]

RAREWORDS = set(RAREWORDS + NUM_WORDS + ADDRESS_WORDS)
print(RAREWORDS)

def remove_rarewords(text):
    """custom function to remove the rare words"""
    return " ".join([word for word in str(text).split() if word not in RAREWORDS])

eng_df["text_wo_stopfreqrare"] = eng_df["text_wo_stopfreq"].apply(lambda text: remove_rarewords(text))
eng_df.head()

def remove_urls(text):
    url_pattern = re.compile(r'https?://\S+|www\.\S+')
    return url_pattern.sub(r'', text)

eng_df["text_wo_stopfreqrare"] = eng_df["text_wo_stopfreqrare"].apply(remove_urls)

def remove_html(text):
    html_pattern = re.compile('<.*?>')
    return html_pattern.sub(r'', text)


eng_df["text_wo_stopfreqrare"] = eng_df["text_wo_stopfreqrare"].apply(remove_html)

eng_df['text_wo_stopfreqrare']

from sklearn.feature_extraction.text import CountVectorizer

from sklearn.model_selection import train_test_split


data_train, data_test = train_test_split(eng_df['text_wo_stopfreqrare'], test_size=0.2, random_state=42)

training_corpus = data_train.values

vectorizer = CountVectorizer(min_df=3)

vectorizer.fit(training_corpus)
num_words = len(vectorizer.vocabulary_)
print("Number of distinct words to be used as features:",num_words)

vectorizer.vocabulary_
