==> eda/9.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../data/process_v1/process_data_5.csv')

lat_lon_df = df[['lat', 'lon']].drop_duplicates()
lat_lon_df

(lat_lon_df['lon'].min(), lat_lon_df['lon'].max())

df[['lat', 'lon', 'district', 'street', 'ward']].to_csv("encode_location.csv", index = False)

import quads

tree = quads.QuadTree((15.86914545, 106.13803025), 12 , 2)

lat_list, lon_list = lat_lon_df['lat'].tolist(), lat_lon_df['lon'].tolist()
lat_lon_tuple_list = [(lat, lon) for lat, lon in zip(lat_list, lon_list)]
print(len(lat_lon_tuple_list))

for lat_lon_tuple in lat_lon_tuple_list:
    tree.insert(lat_lon_tuple)

lat_lon_tuple in tree

found = tree.nearest_neighbors(lat_lon_tuple, count=4)
print(found[0].x, found[0].y)

def get_address_by_lat_lon(lat, lon):
    item_df = df[(df['lat'] == lat) & (df['lon'] == lon)]
    return item_df['district'].mode().tolist()[0], item_df['street'].mode().tolist()[0], item_df['ward'].mode().tolist()[0]

def find_nearest_neighbors(lat, lon, count = 10):
    found = tree.nearest_neighbors((lat, lon), count=count)
    nearest_lat_lon_df = pd.DataFrame()
    nearest_lat_lon_df['lat'] = [item.x for item in found]
    nearest_lat_lon_df['lon'] = [item.y for item in found]

    nearest_lat_lon_df['district'], nearest_lat_lon_df['street'], nearest_lat_lon_df['ward'] = zip(*nearest_lat_lon_df.apply(lambda x: get_address_by_lat_lon(x['lat'], x['lon']), axis = 1))

    arr = np.concatenate(nearest_lat_lon_df.values)
    return arr

tree

lat_lon_df.to_csv('./realestate_lat_lon.csv', index = False)

arrs = [find_nearest_neighbors(lat_lon_tuple[0], lat_lon_tuple[1]) for lat_lon_tuple in lat_lon_tuple_list]
nearest_df = pd.DataFrame(arrs)

nearest_df = pd.DataFrame(arrs)
nearest_df = nearest_df.rename(columns = {0: 'lat', 1: 'lon'})
del nearest_df[2],nearest_df[3], nearest_df[4]
cols = [c for c in nearest_df.columns.tolist() if c not in ['lat', 'lon']]

lat_cols = [c for c in nearest_df.columns.tolist() if c not in ['lat', 'lon'] and c % 5 == 0]
format_lat_cols = [f'nearest_{i}_lat' for i, col in enumerate(lat_cols)]

lon_cols = [c for c in nearest_df.columns.tolist() if c not in ['lat', 'lon'] and c % 5 == 1]
format_lon_cols = [f'nearest_{i}_lon' for i, col in enumerate(lon_cols)]

district_cols = [c for c in nearest_df.columns.tolist() if c not in ['lat', 'lon'] and c % 5 == 2]
format_district_cols = [f'nearest_{i}_district' for i, col in enumerate(district_cols)]

street_cols = [c for c in nearest_df.columns.tolist() if c not in ['lat', 'lon'] and c % 5 == 3]
format_street_cols = [f'nearest_{i}_street' for i, col in enumerate(street_cols)]

ward_cols = [c for c in nearest_df.columns.tolist() if c not in ['lat', 'lon'] and c % 5 == 4]
format_ward_cols = [f'nearest_{i}_ward' for i, col in enumerate(ward_cols)]

source_cols_list = [lat_cols, lon_cols, district_cols, street_cols, ward_cols]
target_cols_list = [format_lat_cols, format_lon_cols, format_district_cols, format_street_cols, format_ward_cols]


for source_cols, target_cols in zip(source_cols_list, target_cols_list):
    rename_dict = dict(zip(source_cols, target_cols))
    nearest_df = nearest_df.rename(columns = rename_dict)
nearest_df

df = df.merge(nearest_df, how='left', on = ['lat', 'lon'])
df

df

import math
from tqdm import tqdm
def distance_func(lat1: float, lon1: float, lat2: float, lon2: float):

    try:
        R = 6371
        dLat = (lat2-lat1) * math.pi / 180
        dLon = (lon2-lon1) * math.pi / 180
        lat1 = lat1 * math.pi / 180
        lat2 = lat2 * math.pi / 180
        a = math.sin(dLat/2) * math.sin(dLat/2) + math.sin(dLon/2) * \
            math.sin(dLon/2) * math.cos(lat1) * math.cos(lat2)
        c = 2 * math.atan2(math.sqrt(a), math.sqrt(1-a))
        d = R * c
        return d*1000
    except:
        print(f"Distance_Func error with: {(lat1, lon1, lat2, lon2)}")


for i in tqdm(range(9)):
    df[f'distance_nearest_{i}'] = df.apply(lambda x: distance_func(x['lat'], x['lon'], x[f'nearest_{i}_lat'], x[f'nearest_{i}_lon']), axis = 1)
    df[f'distance_nearest_{i}'] = np.log(df[f'distance_nearest_{i}'])

df

df.to_csv('../data/process_v1/process_data_6.csv', index = False)

df['time']
